refactor robotInteractionCockpit
