+++ frontend/src/components/pages/RobotInteractionCockpit/RobotInteractionSections/RobotInteractionExecutionSection/singleComponents/RobotLogEntryCard.jsx
