+++ frontend/src/components/pages/RobotInteractionCockpit/RobotInteractionSections/RobotInteractionInputSection/singleComponents/RobotInteractionParameterInput.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Typography, Input, Tooltip } from 'antd';
 import { InfoCircleOutlined } from '@ant-design/icons';
 import PropTypes from 'prop-types';
-import corporateDesign from '../../../../../layout/corporateDesign';
+import corporateDesign from '../../../../../../layout/corporateDesign';
 
 const { Text } = Typography;
 /**
