+++ frontend/src/utils/componentsFunctionality/robotInteractionCockpitFunctionality/robotInteractionCockpitFunctionality.js
